+++ src/media/visual/color/mod.rs
@@ -11,13 +11,10 @@
 mod helpers; // (TEMP Norm)
 
 mod color; // Color
-mod gamma; // GammaConst TEMP
+mod gamma; // Gamma
 mod luminance; // Luma, Luminance
 mod palette; // WIP
 mod rgb; // Rgb[a][8|16|F32|F64], Rgb[a]Lin[F32|F64]
-
-#[cfg(feature = "std")]
-mod gamma_std; // Gamma IMPROVE: merge with GammaConst
 
 // #[cfg(feature = "linear")]
 // mod xyz; // Xyz
@@ -31,8 +28,6 @@ crate::structural_mods! { // _mods, _crate_internals
             palette::_all::*,
             rgb::_all::*,
         };
-        #[cfg(feature = "std")]
-        pub use super::gamma_std::Gamma;
         // #[cfg(feature = "linear")]
         // pub use super::xyz::*;
     }